\#8 Cambios en contexto/gulpfile.js
Ahora nos automatiza el movimiento y de .htacces y wp-config.php cuando aparecen
en /opt/app-roo/src, creándo también sus accesos directos desde
/opt/app-root/src.
                    _ .htaccess-dev (/silo/wordpress/dev/.htaccess-dev)
.htaccess (src) ==>|
                    - .htaccess (/silo/wordpress/.htacces)

                        _ wp-config-dev.php (/silo/wordpress/dev/wp-config-dev.php)
wp-config.php (src) ==>|
                        - .htaccess (/silo/wordpress/wp-config.php)

A su vez con la task "tmovelinkwpcontent" cambiamos el contenido de wp-content a
/silo/wordpress/wp-content.

Todos estos cambios se realizan con los correspondientes accesos directos desde
/opt/app-root/src:

.htacces
wp-content
wp-config.php

+++ contexto/gulpfile.js
@@ -11,10 +11,10 @@ var shell = require('shelljs');
 const srcwpcontent = '/opt/app-root/src/wp-content';
 const srcwpcontentpriv = '/opt/app-root/silo/wordpress/wp-content';
 const dstwpcontent = '/silo/wordpress/wp-content';
-const srcwphtcaccess = '/opt/app-root/src/.htcaccess';
-const srcwphtcaccesspriv = '/opt/app-root/silo/wordpress/.htcaccess';
-const dstwphtcaccessdev = '/silo/wordpress/dev/.htcaccess-dev.php';
-const dstwphtcaccess = '/silo/wordpress/.htcaccess';
+const srcwphtaccess = '/opt/app-root/src/.htaccess';
+const srcwphtaccesspriv = '/opt/app-root/silo/wordpress/.htaccess';
+const dstwphtaccessdev = '/silo/wordpress/dev/.htaccess-dev';
+const dstwphtaccess = '/silo/wordpress/.htaccess';
 const srcwpconfig = '/opt/app-root/src/wp-config.php';
 const srcwpconfigpriv = '/opt/app-root/silo/wordpress/wp-config.php';
 const dstwpconfig = '/silo/wordpress/wp-config.php';
@@ -43,11 +43,14 @@ gulp.task('tmoverwpcontentpriv',async function(done)  {
 gulp.task('tmoverwpcontent',async function(done)  {
   try {
     const existe = await fs.pathExists(srcwpcontent);
-    if ( existe) {
-      await fs.move( srcwpcontent, dstwpcontent, {overwrite: true});
-      console.log('success!');
-    } else {
-      console.log("No existe la carpeta ",srcwpcontent);
+    const existe1 = await fs.pathExists(dstwpcontent);
+    if (! existe1 ){
+      if ( existe) {
+        await fs.move( srcwpcontent, dstwpcontent, {overwrite: true});
+        console.log('success!');
+      } else {
+        console.log("No existe la carpeta ",srcwpcontent);
+      }
     }
   } catch (err) {
     console.error(err);
@@ -59,12 +62,15 @@ gulp.task('tmoverwpcontent',async function(done)  {
 gulp.task('tcreateSymlinkwpcontent', async function(done) {
   try {
     const existe = await fs.pathExists(dstwpcontent);
-    if (existe) {
-      await fs.remove( srcwpcontent );
-      await fs.ensureSymlink( dstwpcontent,  srcwpcontent, 'dir');
-      console.log('success!');
-    } else {
-      console.log("No existe el archivo. ", dstwpcontent);
+    const existe1 = await fs.pathExists(srcwpcontent);
+    if (! existe1) {
+      if (existe) {
+        // await fs.remove( srcwpcontent );
+        await fs.ensureSymlink( dstwpcontent,  srcwpcontent, 'dir');
+        console.log('success!');
+      } else {
+        console.log("No existe el archivo. ", dstwpcontent);
+      }
     }
   } catch (err) {
     console.error(err);
@@ -73,7 +79,7 @@ gulp.task('tcreateSymlinkwpcontent', async function(done) {
 });
 
 
-//////// .htaccess aa
+//////// .htaccess
 
 gulp.task('tmoverwphtaccesspriv',async function(done)  {
   try {
@@ -94,12 +100,16 @@ gulp.task('tmoverwphtaccesspriv',async function(done)  {
 
 gulp.task('tmoverwphtaccess',async function(done)  {
   try {
+    const existe1 = await fs.pathExists(dstwphtaccessdev);
     const existe = await fs.pathExists(srcwphtaccess);
-    if ( existe ){
-      await fs.move( srcwphtaccess, dstwphtaccess, {overwrite: true});
-      console.log('success!');
-    } else {
-      console.log("No existe el archivo ", srcwphtaccess);
+    if (! existe1 ){
+      if (existe) {
+        await fs.copy( srcwphtaccess, dstwphtaccess, {overwrite: true});
+        await fs.move( srcwphtaccess, dstwphtaccessdev, {overwrite: true});
+        console.log('success!');
+      } else {
+        console.log("No existe el archivo ", srcwphtaccess);
+      }
     }
   } catch (err) {
     console.error(err);
@@ -112,12 +122,15 @@ gulp.task('tmoverwphtaccess',async function(done)  {
 gulp.task('tcreateSymlinkwphtaccess', async function(done) {
   try {
     const existe = await fs.pathExists(dstwphtaccessdev);
-    if (existe) {
-      await fs.remove( srcwphtaccess );
-      await fs.ensureSymlink( dstwphtaccessdev,  srcwphtaccess, 'file');
-      console.log('success!');
-    } else {
-      console.log("No existe el archivo ", dstwphtaccessdev);
+    const existe1 = await fs.pathExists(srcwphtaccess);
+    if (! existe1) {
+      if (existe) {
+       //  await fs.remove( srcwphtaccess );
+        await fs.ensureSymlink( dstwphtaccessdev,  srcwphtaccess, 'file');
+        console.log('success!');
+      } else {
+        console.log("No existe el archivo ", dstwphtaccessdev);
+      }
     }
   } catch (err) {
     console.error(err);
@@ -148,8 +161,8 @@ gulp.task('tmoverwpconfigpriv',async function () {
 
 gulp.task('tmoverwpconfig',async function (done) {
   try {
-    const existe = await fs.pathExists(dstwpconfigdev);
-    const existe1 = await fs.pathExists(srcwpconfig);
+    const existe1 = await fs.pathExists(dstwpconfigdev);
+    const existe = await fs.pathExists(srcwpconfig);
     if (! existe1 ) {
       if (existe) {
         await fs.copy( srcwpconfig, dstwpconfig, {overwrite: true});
@@ -169,17 +182,19 @@ gulp.task('tmoverwpconfig',async function (done) {
 gulp.task('twpconfigchange', async function(done) {
   try {
     const existe = await fs.pathExists(dstwpconfigdev);
+    const existe1 = await fs.pathExists(dstwpconfig);
     if ( existe ){
       shell.sed('-i', /^.*define\(.*DB_COLLATE.*/, 'define( \'DB_COLLATE\', \'utf8_general_ci\' );' , dstwpconfigdev);
-      shell.sed('-i', /.*table_prefix.*/, shell.cat ('wp-config.js') , dstwpconfigdev);
-      shell.sed( '-i', /^.*define\(.*WP_DEBUG.*/, shell.cat('wp-config-dev.js') , dstwpconfigdev);
+      shell.sed('-i', /.*table_prefix.*/, shell.cat ('wp_config.js') , dstwpconfigdev);
+      shell.sed( '-i', /^.*define\(.*WP_DEBUG.*/, shell.cat('wp_config_dev.js') , dstwpconfigdev);
+      console.log('success!');
+    }
+    if (existe1 ) {
       shell.sed('-i', /^.*define\(.*DB_HOST.*/, 'define( \'DB_HOST\', \'localhost\' );' , dstwpconfig);
       shell.sed('-i', /^.*define\(.*DB_COLLATE.*/, 'define( \'DB_COLLATE\', \'utf8_general_ci\' );' , dstwpconfig);
-      shell.sed('-i', /.*table_prefix.*/, shell.cat ('wp-config.js') , dstwpconfig);
+      shell.sed('-i', /.*table_prefix.*/, shell.cat ('wp_config.js') , dstwpconfig);
       console.log('success!');
-    } else {
-      console.log("No existe el archivo ", srcwpconfigdev);
-    }
+    } 
   } catch (err) {
     console.error(err);
   };
@@ -217,12 +232,17 @@ gulp.task('tcreateSymlinkwpconfig', async function(done) {
 
 
 gulp.task('watch',  function(done) {
-  gulp.watch('/opt/app-root/src/**/wp-confi?.php',  gulp.series('twpconfigdev','twpconfig','tcreateSymlinkwpconfig'));
-  gulp.watch('/opt/app-root/src/**/.htacces?',  gulp.series('twpconfigdev','twpconfig','tcreateSymlinkwpconfig'));
+  gulp.watch('/opt/app-root/src/**/wp-confi?.php',  gulp.series('tmoverwpconfig','twpconfigchange', 'tcreateSymlinkwpconfig'));
+  gulp.watch('/opt/app-root/src/**/.htacces?',  gulp.series('tmoverwphtaccess', 'tcreateSymlinkwphtaccess'));
+  
   
   done();
 });
 
 gulp.task('default',  gulp.series('watch'));
 
-gulp.task('union', gulp.series('tmoverwpconfig','twpconfigchange', 'tcreateSymlinkwpconfig'));
+// gulp.task('union', gulp.series('tmoverwpconfig','twpconfigchange', 'tcreateSymlinkwpconfig'));
+
+// gulp.task('unionh', gulp.series('tmoverwphtaccess', 'tcreateSymlinkwphtaccess'));
+
+gulp.task('tmovelinkwpcontent', gulp.series('tmoverwpcontent', 'tcreateSymlinkwpcontent'));
